Django Rest Framework/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.8
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
